+++ CHANGES.md
@@ -1,6 +1,8 @@
 unreleased
 ----------
 
+- Support `--chmod` argument in `COPY` and `ADD` Dockerfile
+  instructions. (@MisterDA #174)
 - Add Debian 12 as main distribution. (@MisterDA #172)
 - Add OpenSUSE Leap 15.5 to Tier 2. (@MisterDA #171)
 - Add OpenSUSE Tumbleweed to Tier 2. (@MisterDA #168 #169)